+++ src/plugins/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(input)
 add_subdirectory(output)
+add_subdirectory(process)
