+++ src/plugins/process/basicplus/src/basicplus.hpp
@@ -1,33 +1,16 @@
 /**
- * \file basicplus.hpp
- * \brief Plugin for parsing basicplus traffic.
- * \author Jiri Havranek <havranek@cesnet.cz>
- * \date 2021
- */
-/*
- * Copyright (C) 2021 CESNET
+ * @file
+ * @brief Plugin for parsing basicplus traffic.
+ * @author Jiri Havranek <havranek@cesnet.cz>
+ * @author Pavel Siska <siska@cesnet.cz>
+ * @date 2025
  *
- * LICENSE TERMS
+ * Copyright (c) 2025 CESNET
  *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in
- *    the documentation and/or other materials provided with the
- *    distribution.
- * 3. Neither the name of the Company nor the names of its contributors
- *    may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
- *
- *
- *
+ * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#ifndef IPXP_PROCESS_BASICPLUS_HPP
-#define IPXP_PROCESS_BASICPLUS_HPP
+#pragma once
 
 #include <sstream>
 #include <string>
@@ -40,7 +23,7 @@
 #include <ipfixprobe/flowifc.hpp>
 #include <ipfixprobe/ipfix-elements.hpp>
 #include <ipfixprobe/packet.hpp>
-#include <ipfixprobe/process.hpp>
+#include <ipfixprobe/processPlugin.hpp>
 
 namespace ipxp {
 
@@ -157,7 +140,7 @@ struct RecordExtBASICPLUS : public RecordExt {
  */
 class BASICPLUSPlugin : public ProcessPlugin {
 public:
-	BASICPLUSPlugin();
+	BASICPLUSPlugin(const std::string& params);
 	~BASICPLUSPlugin();
 	void init(const char* params);
 	void close();
@@ -176,4 +159,3 @@ public:
 };
 
 } // namespace ipxp
-#endif /* IPXP_PROCESS_BASICPLUS_HPP */
